+++ apps/frontend/src/_store/redux/rootSaga.ts
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import { watchRecord, watchStartChat, watchRetry, watchCancelCurrentRequestInterview } from "@/store/redux/features/chat/saga";
+import { watchRecord, watchStartChat, watchRetry, watchCancelCurrentRequestInterview } from "@/_store/redux/features/chat/saga";
 
 export function* helloSaga() {
   console.log("Hello Sagas!");
