+++ apps/frontend/src/_store/redux/type.ts
