+++ apps/frontend/src/app/_components/Header/Header.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import Header from "./Header";
-import useUserStore from "@/store/useUserStore";
+import useUserStore from "@/_store/useUserStore";
 
 const mockUser = {
   id: 1,
