+++ apps/frontend/src/app/chat/[name]/_components/ChatArticle.tsx
@@ -3,14 +3,14 @@ import styles from "./chat.module.css";
 import Avatar from "@repo/ui/Avatar";
 import clsx from "clsx";
 import Button from "@repo/ui/Button";
-import { ChatContentStatusType } from "@/store/redux/type";
+import { ChatContentStatusType } from "@/_store/redux/type";
 import { useDispatch, useSelector } from "react-redux";
 import {
   REQUEST_INTERVIEW,
   RETRY_INTERVIEW,
   CANCEL_CURRENT_REQUEST_INTERVIEW,
-} from "@/store/redux/features/chat/slice";
-import { selectChatId } from "@/store/redux/features/chat/selector";
+} from "@/_store/redux/features/chat/slice";
+import { selectChatId } from "@/_store/redux/features/chat/selector";
 
 export type ChatType = "user" | "bot" | "";
