+++ apps/frontend/src/app/chat/[name]/page.tsx
@@ -4,8 +4,8 @@ import InterviewerProfileSection from "./_components/InterviewerProfileSection";
 import styles from "./page.module.css";
 import Text from "@repo/ui/Text";
 import ChatSection from "./_components/ChatSection";
-import useUserStore from "@/store/useUserStore";
-import { useInterviewerStore } from "@/store/useInterviewerStore";
+import useUserStore from "@/_store/useUserStore";
+import { useInterviewerStore } from "@/_store/useInterviewerStore";
 import { useRouter } from "next/navigation";
 
 export default function Page() {
